fix(platform-clipboard): move rep_bytes to payload.rs so Linux can compile
CI failure on PR #772: Linux Rust Check + coverage failed with
`cannot find 'common' in 'clipboard'` because `clipboard::common` is
gated to `#[cfg(any(target_os = "macos", target_os = "windows"))]`
(Phase 4 narrowed clipboard-rs to those targets). The Linux Wayland /
X11 selection writers tried to call `crate::clipboard::common::rep_bytes`
which doesn't exist on Linux.

Extract `rep_bytes` to a new `clipboard::payload` module that has no
clipboard-rs dependency, so all three platforms can import it. The
regression tests move with it (they no longer have to live in macos.rs
to be exercised, and now run on Linux CI too).

- New `clipboard/payload.rs` owns `rep_bytes` + its 3 tests.
- `clipboard/mod.rs` declares `pub(crate) mod payload;` without cfg.
- `clipboard/common.rs` drops the local copy, re-uses
  `super::payload::rep_bytes` via plain `use`.
- macos.rs / windows.rs / linux/{wlr,ext,x11/writer}.rs all switch to
  `crate::clipboard::payload::rep_bytes`.
- macos.rs tests module drops the now-duplicate rep_bytes tests.

cargo check passes on macOS, x86_64-pc-windows-gnu; 32 lib tests green
on macOS including the 3 relocated `payload::tests::rep_bytes_*` cases.
