+++ src-tauri/crates/uc-platform/src/clipboard/platform/linux/wayland/protocol/wlr.rs
@@ -689,9 +689,9 @@ fn handle_write(
             // 用 `rep_bytes` 而非 `expect_inline_bytes`：远端 push 的 image rep 由
             // `apply_inbound::materializer` 合成为 `LocalFile` source（指向 blob cache
             // 文件），直接调 `expect_inline_bytes` 会 panic（见
-            // `clipboard::common::rep_bytes` 注释）。读盘失败时跳过该 rep + warn,
+            // `clipboard::payload::rep_bytes` 注释）。读盘失败时跳过该 rep + warn,
             // 与 macOS / Windows 平台同语义。
-            let bytes = match crate::clipboard::common::rep_bytes(rep) {
+            let bytes = match crate::clipboard::payload::rep_bytes(rep) {
                 Ok(b) => b.into_owned(),
                 Err(err) => {
                     warn!(
