+++ src-tauri/crates/uc-platform/src/clipboard/platform/linux/x11/writer.rs
@@ -119,9 +119,9 @@ pub(super) fn install_snapshot(
         // 用 `rep_bytes` 而非 `expect_inline_bytes`：远端 push 的 image rep 由
         // `apply_inbound::materializer` 合成为 `LocalFile` source（指向 blob cache
         // 文件），直接调 `expect_inline_bytes` 会 panic（见
-        // `clipboard::common::rep_bytes` 注释）。读盘失败时跳过该 rep + warn,
+        // `clipboard::payload::rep_bytes` 注释）。读盘失败时跳过该 rep + warn,
         // 与 macOS / Windows / Wayland 平台同语义。
-        let bytes = match crate::clipboard::common::rep_bytes(rep) {
+        let bytes = match crate::clipboard::payload::rep_bytes(rep) {
             Ok(b) => Arc::new(b.into_owned()),
             Err(err) => {
                 warn!(
